Fix a boo-boo in the .bbclass...
- Technically it's wrong.  It's in the bbclass, so we should be
  checking for an empty dir there for things on the services we're
  automating for.  That being said?  You're NOT supposed to be
  doing an empty for anything using the .bbclass here.
- So, we're going to let it silently fail to package right for now
  until I can get some of the downstream mods in for this (Hopefully
  shortly now...) and then we can trap for the failure there.  We
  should fatal error on this to be "correct" with the full behavior.

+++ classes/runit.bbclass
@@ -85,6 +85,8 @@ enable_default_services() {
 
     cd ${D}${runit-svcdir}
     for svc in *; do
+        # Catch situations where we don't have ANY files (* comes back for the globbing which is broken for this.)
+        [ -f "$svc" ] || continue
         svc="$(basename $svc)"
         ln -s ${runit-svcdir}/$svc ${D}${runit-runsvdir}/default
         for option in $options ; do
